Improvements / additions to AE discard logic
- Move DiscardCondition enum and its parser to AttachEffectTypeClass files
- Introduce a new public field in AttachEffectClass that can be set to true to discard AE (useful for externally handled discard condition checks)
- Add discard condition for weapon firing
