+++ src/Utilities/Enum.h
@@ -214,19 +214,6 @@ enum class ChronoSparkleDisplayPosition : unsigned char
 
 MAKE_ENUM_FLAGS(ChronoSparkleDisplayPosition);
 
-enum class DiscardCondition : unsigned char
-{
-	None = 0x0,
-	Entry = 0x1,
-	Move = 0x2,
-	Stationary = 0x4,
-	Drain = 0x8,
-	InRange = 0x10,
-	OutOfRange = 0x20
-};
-
-MAKE_ENUM_FLAGS(DiscardCondition);
-
 enum class ExpireWeaponCondition : unsigned char
 {
 	None = 0x0,
